+++ lib/chef/resource/aws_launch_configuration.rb
@@ -3,9 +3,6 @@ require 'chef/provisioning/aws_driver/aws_resource'
 class Chef::Resource::AwsLaunchConfiguration < Chef::Provisioning::AWSDriver::AWSResource
   aws_sdk_type AWS::AutoScaling::LaunchConfiguration, id: :name
 
-  actions :create, :delete, :nothing
-  default_action :create
-
   attribute :name,          kind_of: String, name_attribute: true
   attribute :image,         kind_of: [ String, AWS::EC2::Image ]
   attribute :instance_type, kind_of: String