+++ lib/chef/resource/aws_load_balancer.rb
@@ -3,9 +3,6 @@ require 'chef/provisioning/aws_driver/aws_resource'
 class Chef::Resource::AwsLoadBalancer < Chef::Provisioning::AWSDriver::AWSResource
   aws_sdk_type AWS::ELB::LoadBalancer, load_provider: false
 
-  actions :create, :delete, :nothing
-  default_action :create
-
   attribute :name, kind_of: String,  name_attribute: true
 
   attribute :load_balancer_id, kind_of: String, aws_id_attribute: true, lazy_default: proc {